requirements.txt
streamlit
pdfplumber
openai
sentence-transformers
numpy
torch
python-dotenv
googlesearch-python